backend/requirements.txt
# ── Web Framework ──────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# ── OCR ────────────────────────────────────────────────────────
easyocr==1.7.1

# ── Image Processing ───────────────────────────────────────────
opencv-python-headless==4.9.0.80
numpy==1.26.4
Pillow==10.3.0

# ── PDF Support ────────────────────────────────────────────────
pdf2image==1.17.0
# Note: poppler required on system → added in render.yaml buildCommand

# ── LLM ────────────────────────────────────────────────────────
openai==1.30.1

# ── Utilities ──────────────────────────────────────────────────
python-dotenv==1.0.1